arreglo codigo barras de vida

+++ Assets/Scripts/BarraDeVidaBoss.cs
@@ -1,4 +1,3 @@
-using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -15,17 +14,17 @@ public class BarraDeVidaBoss : MonoBehaviour
     {
         _Boss = GameObject.Find("boss").GetComponent<FinalBoss>();
         vidaMaxima = _Boss.lives;
-        
 
     }
     // Update is called once per frame
     void Update()
     {
-        barraDeVidaBoss.fillAmount = _Boss.lives / vidaMaxima;
-        if(_Boss._alive == false)
+        barraDeVidaBoss.fillAmount = (float)_Boss.lives / vidaMaxima;
+        if (_Boss._alive == false)
         {
             Destroy(gameObject);
-            
+
         }
     }
 }
+
